+++ packages/suite-web-landing/babel.config.js
@@ -10,6 +10,7 @@ module.exports = api => {
                     '^@suite-web-landing-components': '../../packages/suite-web-landing/components/\\1/index',
                     '^@landing-components/(.+)': '../../packages/landing-page/components/\\1/\\2',
                     '^@landing-components': '../../packages/landing-page/components/\\1/index',
+                    '^@(.+)-components/(.+)': '../../packages/suite/src/components/\\1/\\2',
                     '^@(.+)-config/(.+)': '../../packages/suite/src/config/\\1/\\2',
                     '^@(.+)-config': '../../packages/suite/src/config/\\1/index',
                     '^@(.+)-constants/(.+)': '../../packages/suite/src/constants/\\1/\\2',