+++ packages/suite-web-landing/pages/index.tsx
@@ -5,9 +5,10 @@ import Layout from '@suite-web-landing-components/Layout';
 import Translation, { TranslationModeContext } from '@suite-web-landing-components/Translation';
 import Download from '@suite-web-landing-components/Download';
 import Feature from '@suite-web-landing-components/Feature';
-import Metadata from '@suite-web-landing-components/Metadata';
 import { resolveStaticPath } from '@suite-utils/nextjs';
 import { H1, P, variables, colors } from '@trezor/components';
+import Metadata from '@suite-components/Metadata';
+import { URLS } from '@suite-constants';
 import { Fade } from 'react-awesome-reveal';
 import enLocale from '@trezor/suite-data/files/translations/en.json';
 
@@ -136,7 +137,11 @@ const Index = () => {
     return (
         <TranslationModeContext.Provider value={translationMode}>
             <IntlProvider locale="en" messages={enLocale}>
-                <Metadata />
+                <Metadata
+                    image={`${URLS.SUITE_URL}${resolveStaticPath(
+                        'images/suite-web-landing/meta.png',
+                    )}`}
+                />
                 <Layout>
                     <Wrapper>
                         <StyledHeroCta>