+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Discovery progress bar in electron
+- OpenGraph meta tags for Suite (web version)
 
 ### Changed